refactor(validation): modularize traceability-validation.js (993 → 9 modules)
SD-LEO-REFACTOR-TRACEABILITY-001 - Child 10 of orchestrator SD-LEO-REFACTOR-LARGE-FILES-002

Extracted 993 LOC Gate 3 (PLAN→LEAD) validation into focused modules:

Utils (1 file):
- utils.js - Path resolution, SD context resolution

Preflight (1 file):
- preflight/index.js - Phase 1 non-negotiable blockers

Sections (6 files):
- sections/recommendation-adherence.js - Section A (30 pts CRITICAL)
- sections/implementation-quality.js - Section B (30 pts CRITICAL)
- sections/traceability-mapping.js - Section C (25 pts MAJOR)
- sections/sub-agent-effectiveness.js - Section D (10 pts MINOR)
- sections/lessons-captured.js - Section E (5 pts MINOR)
- sections/index.js - Section exports

Main (1 file):
- index.js - Main orchestrator with validateGate3PlanToLead

Original file now thin re-export (17 LOC). 11 exports verified.
