+++ scripts/modules/traceability-validation/sections/index.js
@@ -1,0 +1,10 @@
+/**
+ * Section Exports for Traceability Validation
+ * Part of SD-LEO-REFACTOR-TRACEABILITY-001
+ */
+
+export { validateRecommendationAdherence } from './recommendation-adherence.js';
+export { validateImplementationQuality } from './implementation-quality.js';
+export { validateTraceabilityMapping } from './traceability-mapping.js';
+export { validateSubAgentEffectiveness } from './sub-agent-effectiveness.js';
+export { validateLessonsCaptured } from './lessons-captured.js';
